pyproject.toml
[project]
name = "m26003project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "requests>=2.33.1",
]
